==> src/1_general_cleaning.ipynb <==
# SHARKNADO

import pandas as pd
pd.set_option('display.max_columns', None)
#pd.set_option('display.max_rows',None)

import numpy as np
import re

import warnings
warnings.filterwarnings('ignore')

df = pd.read_csv("../data/0_attacks.csv", encoding='latin-1')

# 1. Limpieza de filas nulas (aproximación general)
Lo primero comprobamos si existen filas donde todos los valores sean nulos.

len(df[df.isna().all(axis=1)])

df = df.dropna(how='all')

Eliminamos las 17020 filas donde todos los valores son nulos a su falta de aportación para el Data Frame.

Ahora filtramos por filas que tengan menos de 1 o 2 valores. Observamos que son las mismas que aquellas que tienen menos de 10 valores por lo tanto decidimos eliminarlas debido a la falta de interés en las mismas. 

result = df[df.count(axis=1) < 3]
len(result)

result = df[df.count(axis=1) < 10]
len(result)

(df[df.count(axis=1) < 3]).equals(df[df.count(axis=1) < 10])

df = df[~(df.count(axis=1) < 10)]

Tras eliminar las columnas con menos de 10 valores efectivos nos queda un Data Frame de 6302 registros con los que podemos trabajar y que podrían tener interés. A continuación se procederá a abordar el data frame columna por columna, revisando los valores y tratando de despejar los nan.

# 2. Limpieza de columnas nulas (aproximación general)

df.head(0)

Tras revisar las variables de las columnas empezamos la limpieza por las dos últimas columnas. A continuación comprobamos el contenido de las columnas

df["Unnamed: 22"].unique()

df["Unnamed: 23"].unique()

Vemos que los únicos datos que ofrecen estas columnas son: [nan, 'Teramo', 'stopped here', 'change filename']. Revisaremos las veces que se repiten los datos para comprobar su utilidad, aunque a primera vista parecen columnas inservibles.

df["Unnamed: 22"].value_counts()

df["Unnamed: 23"].value_counts()

df = df.drop(["Unnamed: 22", "Unnamed: 23"], axis = 1)

df.shape

df.to_csv("../data/1_shark_limpio.csv", sep=',', index=False)
