==> src/2_date_cleaning.ipynb <==
import pandas as pd
pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows',None)

from datetime import datetime
import numpy as np
import re

import warnings
warnings.filterwarnings('ignore')

df = pd.read_csv("../data/1_shark_limpio.csv", encoding='latin-1')

df.columns = df.columns.str.strip()

df = df.drop(["Sex","Species" ,"Age", "Type", "Country", "Area", "Location", "Activity", "Name", "Injury", "Fatal (Y/N)","Time", "Investigator or Source","pdf","href formula","href","original order"],axis = 1)

# Limpieza de fecha

df.head()

def sacar_fecha(date_str):
    # Define a regex pattern to match the components with different delimiters
    pattern = r'(\d{4})[/.-](\d{2})[/.-](\d{2})'

    # Use the regex pattern to match and extract the components
    matches = re.search(pattern, date_str)

    if matches:
        year = matches.group(1)
        month = matches.group(2)
        day = matches.group(3)
        return year, month, day
    else:
        return 0, 0, 0

df[['Y_casen1', 'M_casen1', 'D_casen1']] = df['Case Number.1'].apply(sacar_fecha).apply(pd.Series)

df[['Y_casen2', 'M_casen2', 'D_casen2']] = df['Case Number.2'].apply(sacar_fecha).apply(pd.Series)

#df = df.drop(["Case Number.1", "Case Number.2"], axis=1)

df.head()

Establecemos las funciones para de la columna date sacar año, mes y día.

def extract_year(date_string):
    pattern = r"(\d{4})"
    match = re.search(pattern, date_string)
    if match:
        return match.group(1)
    else:
        return "0"


# Apply the extract_year function to the Date column
df['Year'] = df['Date'].apply(extract_year)
df['Year'] = df['Year'].astype(int)

def extract_month(date_string):
    match = re.search(r'(\d{2}-[A-Za-z]{3}-\d{4})', date_string)
    if match:
        date_part = match.group(1)
        date_object = datetime.strptime(date_part, "%d-%b-%Y")
        return date_object.month
    else:
        return "0"

# Apply the function to the 'Date' column
df['Month'] = df['Date'].apply(extract_month)
df['Month'] = df['Month'].astype(int)

def extract_day(date_string):
    pattern = r"(\d{2})"
    match = re.search(pattern, date_string)
    if match:
        return match.group(1)
    else:
        return "0"


# Apply the extract_year function to the Date column
df['Day'] = df['Date'].apply(extract_day)
df['Day'] = df['Day'].astype(int)

#df = df.drop(["Date"], axis = 1)

Ahora tenemos tres columnas en las que presentamos los datos de la fecha. A continuación desecharemos la columna "Date" ya que no tiene sentido conservarla. Despues revisaremos los valores creados

df.info()

#df[df.isna().any(axis=1)]

Vemos que solo existe un NaN en el Data Frame en la columna Case Number. Observando la fila vemos que podemos rellenar este valor con los datos de Case Number.1 o Case Number.2

df['Case Number'].fillna(df['Case Number.1'], inplace=True)

df.loc[5488]

A continuación pasamos a revisar los datos creados para verificar su eficacia

Vemos que hay 19 valores donde el año es 0. Tratamos de sobreescribirlos con la información en el resto de columnas.

df[df["Year"] == 0]

df.at[4644, 'Year'] = 1944
df.at[4802, 'Year'] = 1939
df.at[4838, 'Year'] = 1937

df.at[4644, 'Month'] = 7
df.at[4802, 'Month'] = 0
df.at[4838, 'Month'] = 9

df.at[4644, 'Day'] = 22
df.at[4802, 'Day'] = 0
df.at[4838, 'Day'] = 26

#indices = df[df['Final_year'] == 0].index.tolist()
#indices

#df = df.drop(indices)

Nos sacamos valores que no interesan. Previos a la IIWW. Inconclusos o sin suficientes datos. A continuación revisamos dias y años

# Checkpoint

Ahora que hemos creado nuevas columnas y tenemos varios datos buscaremos cruzar las tablas para unificar el df

df.shape

df.head(0)

df[['Y_casen1', 'M_casen1', 'D_casen1']]=df[['Y_casen1', 'M_casen1', 'D_casen1']].astype(int)

df[['Y_casen2', 'M_casen2', 'D_casen2']]=df[['Y_casen2', 'M_casen2', 'D_casen2']].astype(int)

# Function to apply the conditions
def apply_conditions(row):
    original = row['Year']
    casen1 = row['Y_casen1']
    casen2 = row['Y_casen2']

    if original == casen1:
        return casen1
    elif original == casen2:
        return casen2
    else:
        return original

# Apply the function to create a final 'yaer' column
df['Final_year'] = df.apply(apply_conditions, axis=1)

# Function to apply the conditions
def apply_conditions(row):
    original = row['Month']
    casen1 = row['M_casen1']
    casen2 = row['M_casen2']

    if original == casen1:
        return casen1
    elif original == casen2:
        return casen2
    else:
        return original

# Apply the function to create a final 'yaer' column
df['Final_month'] = df.apply(apply_conditions, axis=1)

# Function to apply the conditions
def apply_conditions(row):
    original = row['Day']
    casen1 = row['D_casen1']
    casen2 = row['D_casen2']

    if original == casen1:
        return casen1
    elif original == casen2:
        return casen2
    else:
        return original

# Apply the function to create a final 'yaer' column
df['Final_day'] = df.apply(apply_conditions, axis=1)

df.head(0)

df_fechas = df[["Final_year", "Final_month", "Final_day"]]

df_fechas.shape

df_fechas.to_csv("../data/2_shark_fechas.csv")
