==> src/3_Index_Type_Cleaning.ipynb <==
import pandas as pd
pd.set_option('display.max_columns', None)
#pd.set_option('display.max_rows',None)

import numpy as np
import re

import warnings
warnings.filterwarnings('ignore')

df_clean = pd.read_csv("../data/1_shark_limpio.csv", encoding='latin-1')

df_date = pd.read_csv("../data/2_shark_fechas.csv", encoding='latin-1', sep = ",")

Concatenamos ambos df y eliminamos las columnas sobrantes ya que ahora tenemos la fecha desglosada en las columnas Final_XXXX.

Además, eliminamos las filas en las que tras la esploración no se ha logrado determinar fecha

df = pd.concat([df_clean, df_date], axis = 1)

df = df.drop(["Date", "Case Number.1", "Case Number.2", "Unnamed: 0"], axis=1)

Ahora que el case number ya no tiene sentido mantenerlo lo vamos a reconvertir en el índice para los ataques. Sin embargo antes vamos a eliminar filas que tras la exploración no se hayan podido corregir

list(df[df['Final_year'] == 0]) == list(df[df['Final_month'] == 0])

Vemos que las filas en las que el año final y el mes final que tienen valores de 0 son coincidentes. Por lo tanto decicimos eliminarlas debido a la falta de concreción de una fecha del ataque.

indices = df[df['Final_year'] == 0].index.tolist()

# Eliminar las filas con los índices especificados
df = df.drop(indices)


continuación pasamos la columna de case number para ser un índice propio

df["Case Number"] = ["Attack-" +str(i).zfill(4) for i in range(len(df), 0, -1)]

También eliminamos la columna Year ya que los datos no estaban correctamente presentados y la hemos sustituido por Final_year

df = df.drop(["Year"], axis=1)

Corregimos la siguiente columna. Type. Unificando los valores mal escritos o que significan lo mismo y ocupandonos de los posibles Nan

df = df.replace(["Boat", "Boatomg"], "Boating")

Tras una primera corrección revisamos las dos filas categorizadas como questionable. Se decide eliminar dichas filas debido a su falta de interés para la base de datos. No se puede confirmar la presencia de un tiburon en el ataque y no se presentan mayores daños que desperfectos en una tabla de surf.

df[df.Type == "Questionable"]

df = df.drop(df[df.Type == "Questionable"].index)

A continuación reseteamos el índice y nos ocuparemos de los valores nulos

df = df.reset_index(drop=True)

df[df.Type.isnull()]

Observamos que en la columna solamente hay 4 valores nulos. Sin embargo se decide eliminar las filas debido a la presencia de nulos en otras columnas clave como la especie,

df = df.dropna(subset=['Type'])
df[df.Type.isnull()]

df.Type.value_counts()

df.to_csv("../data/3_shark_index_filas.csv",index=False)
