==> src/4_geography.ipynb <==
import pandas as pd
import numpy as np
import geonamescache
import pycountry
import difflib
import warnings
import re

from countryinfo import CountryInfo
warnings.filterwarnings('ignore')
pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows',None)

df = pd.read_csv("../data/3_shark_index_filas.csv")

df.head()

Ahora que tenemos nuestro propio índice y la columna Type corrediga pasamos, en este notebook, a enfrentarnos a las columnas Country y Area. El objetivo es unificar lo máximo posible a través de la corrección de erratas y nulos. Además también se buscará crear una nueva columna para clasificar los ataques por continente

Empezaremos con una exploración de los datos para luego pasar a la corrección de los mismos.

def corregir_geo(nombre_pais):
    try:
        # Intenta encontrar una coincidencia exacta en pycountry
        pais = pycountry.countries.get(name=nombre_pais)
        if pais:
            return pais.name  # Si se encuentra una coincidencia exacta, devuelve el nombre estándar
        else:
            # Si no se encuentra una coincidencia exacta, intenta encontrar una coincidencia cercana usando difflib
            nombres_estandar = [country.name for country in pycountry.countries]
            coincidencias_cercanas = difflib.get_close_matches(nombre_pais, nombres_estandar, n=1)
            if coincidencias_cercanas:
                return coincidencias_cercanas[0]  # Devuelve la coincidencia más cercana encontrada
            else:
                return nombre_pais  # Si no se encuentra ninguna coincidencia, devuelve el nombre original
    except LookupError:
        return nombre_pais  # Devuelve el nombre original en caso de errores

Vemos que varias filas que tienen nulos en su categoría Country también tienen en Area y Location. Decidimos eliminarlas debido a la falta de concreción geográfica del ataque.

len(df.Country.unique())

df.Country.replace("USA","United States", inplace=True)

df.Country = df.Country.apply(corregir_geo)

df = df.dropna(subset=["Country"])

len(df.Country.unique())

Ahora que se ha corregido la columna Country pasamos a hacer lo mismo para las columnas Area y Location.

print("Before:" ,len(df.Area.unique()))
print("After:", len(df.Location.unique()))

df.Area = df.Area.apply(corregir_geo)
df.Location = df.Location.apply(corregir_geo)

print("Before:", len(df.Area.unique()))
print("After:", len(df.Location.unique()))

Vemos que la función tiene gran éxito corrigiendo las columnas Area y Location, unificando casi 150 y 330 categorias, respectivamente. A continuación veremos los nulos de estas columnas.

len(df[df.Area.isnull()])

len(df[df.Location.isnull()])

Vemos que hay demasiados registros cómo para analizarlos a mano. Trataremos de arreglar la columna Area a través de la localización. A continuación revisamos las columnas donde "Area" sea null y "Location" contenga un elemento.

df_area = df[df.Area.isnull() & ~df.Location.isnull()]

len(df_area.Location)

La longitud continua siendo demasiado grande como para poder determinar el Area a través del Location. Por tanto decidimos asignar el valor "Unknown" a los elementos nulos

df['Location'].fillna("Unknown", inplace=True)

df['Area'].fillna("Unknown", inplace=True)

df.to_csv("../data/4_shark_geography.csv",index=False)
