==> src/5_person.ipynb <==
import pandas as pd
import numpy as np
import geonamescache
import pycountry
import difflib
import warnings
import statistics as stat
import re

from countryinfo import CountryInfo
warnings.filterwarnings('ignore')
pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows',None)

df = pd.read_csv("../data/4_shark_geography.csv")

df.head()

# Limpieza columna "Activity"

def activity(x):
    try:
        x = x.lower()
        if "fis" in x:
            return "Fishing"
        elif "sta" in x:
            return "Standing"
        elif "pad" in x:
            return "Padding"
        elif "sur" in x:
            return "Surfing"
        elif "swi" in x:
            return "Swimming"
        elif "bat" in x:
            return "Bathing"
        elif "div" in x:
            return "Diving"
        elif "sno" in x:
            return "Snorkel"
        elif "boa" in x:
            return "boarding"
        else:
            return "Other"
    except:
        return "Other"

len(df[df.Activity.isnull()])

df.Activity = df.Activity.apply(activity)

df.Activity.value_counts()

len(df[df.Activity.isnull()])

Con la función utilizada simplificamos la columna en 10 categorías. Previamente a la aplicación de la función contamos con 500 nulos que al no tener información sobre la actividad introducimos en la categoría "Otros", la cual resulta ser la más amplia. De cerca le siguen las categorias de pesca y surf con valores muy similares.

# limpieza columna "Name"

A continuación limpiarémos la columna name. Revisando a través de value_counts() detectamos varios valores como male or boy, los cuales nos sirven para verificar la columna de género. Se procede a la utilización y limpieza de estos valores. A continuación, al ser útiles en la columna de nombre los sustituirémos por Unknown, esto serña tras la verificación del género.

def name_1 (x):
    try:
        x = x.lower()
        if "men" in x:
            return "M"
        if "male" in x:
            return "M"
        elif "female" in x:
            return "F"
        elif "girl" in x:
            return "F"
        elif "woman" in x:
            return "F"
        elif "boy" in x:
            return "M"
        elif "man" in x:
            return "M"
        else:
            return x
    except:
        return x
    
def name_2(x):
    if x == "boat":
        return "Unknown"
    elif x == "child":
        return "Unknown"
    elif x== "anonymous":
        return "Unknown"
    elif x == "sailor":
        return "Unknown"
    elif x == "unidentified":
        return "Unknown"
    elif x == "a sailor":
        return "Unknown"
    elif x == "a native":
        return "Unknown"
    elif x == "a soldier":
        return "Unknown"
    elif x == "a pearl diver":
        return "Unknown"
    else:
        return x
    
def name_3(x):
    if x == "M":
        return "Unknown male"
    elif x == "F":
        return "Unknown female"
    elif x == "unknown":
        return "Unknown"
    else:
        return x

df.Name = df.Name.apply(name_1)
df.Name = df.Name.apply(name_2)

Ahora que tenemos los nombres agrupados procedemos a utilizar las categorias creadas "M" y "F" para verificar y rellenar la columna género.

df["Sex"] = df['Sex ']
df["Species"] = df['Species ']
df.drop(["Sex ", "Species "], axis=1, inplace=True)

df.insert(7, 'Sex', df.pop("Sex"))

Vemos 190 registros en los que Name "M"/"F" y Sex son diferentes. Vamos a priorizar el registro dado en la columna Sex, sin embargo tb vemos algunos valores NaN que rellenaremos con el valor obtenido de la exploración previa.

remplazo = df[(df['Name'] == "M") & (df['Sex'] != "M") & df['Sex'].isna()]
df.loc[remplazo.index, 'Sex'] = remplazo['Name']

Para terminar con la columna de nombre vamos a substituir los valores creados M y F por "Unknown", a esta categoría también añadiremos los Nan

df.Name = df.Name.apply(name_3)
df['Name'] = df['Name'].fillna("Unknown")

# Limpieza columna "Sex"

Al ya haber concatenado la columna Sex con los registros obtenidos de la columna Name en este paso únicamente rellenaremos los NaN Sex con valores Unknown. También limpiaremos las categorias desconocidas.

def genero(x):
    if x == "M ":
        return "M"
    elif x == "lli":
        return "M"
    elif x == "N":
        return "Unknown"
    elif x == ".":
        return "Unknown"
    else:
        return x

df.Sex = df.Sex.fillna("Unknown")
df.Sex = df.Sex.apply(genero)

df.Sex.value_counts()

# Limpieza columna "Age"

def age_teen(x):
    try:
        x = x.lower()
        if "teen" in x:
            return 18
        else: return x
    except:
        return x

def age_several(x):
    try:
        nums = re.findall(r'\b\d{2}\b', x)
        nums = [int(num) for num in nums]
        if len(nums) > 0:
            return stat.mean(nums)
        return x
    except: return x

def age_letter(x):
    try:
        x = int(x)
        # Use a regular expression to capture numeric parts
        match = re.search(r'(\d+(\.\d+)?)', x)

        if match:
            num = match.group(1)
            num = re.sub(r'[^\d.]', '', num)  # Remove non-numeric characters except for decimal points
            if "." in num:
                num = int(float(num))  # Convert to float and round to an integer
            return num
        else:
            return x
    except:
        return x

def rellenar(x):
    if type(x) == int:
        return x
    else:
        return 0

df.Age = df.Age.apply(age_teen)
df.Age = df.Age.apply(age_several)
df.Age = df.Age.apply(age_letter)
df.Age = df.Age.apply(rellenar)
df['Age'].fillna(0, inplace=True)  # rellenamos los valores nulos con 0

Para limpiar la columna age se decide recuperar las edades de los strings y convertir la columna a tipo int. Tanto los valores nulos cómo los no deseados se convierten a 0. A la hora de realizar un análisis estadístico de la columna será necesario tener en cuenta que los valores 0 no deben influenciar los resultados.

# Limpieza columna "Injury"

En un primer vistazo a los datos aportados por la columna "Injury" observamos un considerable número de registros catalogados como "FATAL". Decidimos utilizar estos para contrastar los datos aportados por la siguiente columna Fatal (Y/N)

df.Injury.value_counts()[:5]

len(df[(df['Injury'] == 'FATAL') & (df['Fatal (Y/N)'] == 'Y')])

Observamos que hay una diferencia entre el número de registros "FATAL" y el número de coincidencias con la columna fatal(Y/N). A continuación cambiaremos el nombre de la columna fatal para una mayor comodidad. También decidimos que allí donde se haya indicado que la herida fue fatal situaremos el mismo registro en la columna "Deadly"

df["Deadly"] = df['Fatal (Y/N)']
df.drop(['Fatal (Y/N)'],axis=1,inplace=True)

df.loc[df['Injury'] == 'FATAL', 'Deadly'] = 'Y'

len(df[(df['Injury'] == 'FATAL') & (df['Deadly'] == 'Y')])

Vemos que en la columna Injury tenemos 26 nulos. Categorizaremos estos registros como "Unknown". Para finar el tratamiento de estos datos aplicaremos formulas para corregir facilitar la lectura.

len(df[df.Injury.isnull()])

df.Injury.fillna("Unknown",inplace=True)

df.Injury = df.Injury.apply(lambda s: s.lower())
df.Injury = df.Injury.apply(lambda s: s.capitalize())

# Limpieza columna "Deadly"

A continuación limpiamos los valores incongruentes de la columna Deadly categorizandolos como "Unknown".

def dead(x):
    try:
        if x == "N":
            return x
        elif x == "Y":
            return x
        else:
            return "Unknown"
    except:
        return x

df.Deadly = df.Deadly.apply(dead)

df.Deadly.value_counts()

# Limpieza columna "href"

Para la limpieza de las columnas "href" y "href formula" hemos comprobado que únicamente existe un valor nulo en la columna "href formula". También se ha comprobado que aunque aparentemente parezcan contener la misma información existen 60 registros donde estas columnas son diferentes. Debido a la inexistencia de nulos en la columna "href" se decide confiar en esta y eliminar "href formula", ya que no tendría sentido tener dos columnas quee tuvieran la misma información.

len(df[df.href.isnull()])

len(df[df["href formula"].isnull()])

len(df[df.href != df["href formula"]])

df.drop(["href formula"],axis=1,inplace=True)

df.to_csv("../data/5_shark_person.csv",index=False)
