==> src/6_reorganization.ipynb <==
import pandas as pd
import numpy as np
import geonamescache
import pycountry
import difflib
import warnings
import statistics as stat
import re

from countryinfo import CountryInfo
warnings.filterwarnings('ignore')
pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows',None)

Aprovechamos este checkpoint del tratamiento del data frame para reorganizar las columnas y sus nombres.

df = pd.read_csv("../data/5_shark_person.csv")

df["Year"] = df.Final_year
df["Month"] = df.Final_month
df["Day"] = df.Final_day
df["Source"] = df["Investigator or Source"]
df["Season"] = 0
df["Daytime"] = 0
df["Size"] = 0




df = df[["Case Number", "Year", "Month", "Day", "Season", "Time", "Daytime", "Country", "Area", "Location", "Type", "Activity", "Injury", "Deadly", "Species", "Size", "Name", "Sex", "Age", "Source", "pdf", "href", "original order"]]



df.head()

df.shape

df.to_csv("../data/6_shark_reorganization.csv",index=False)
