==> src/7_time.ipynb <==
import pandas as pd
import numpy as np
import geonamescache
import pycountry
import difflib
import warnings
import statistics as stat
import re

from countryinfo import CountryInfo
warnings.filterwarnings('ignore')
pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows',None)

df = pd.read_csv("../data/6_shark_reorganization.csv")
df.head()

# Limpieza columna "Time"

def hour_letter (x):
    try:
        x = x.lower()
        if "nig" in x:
            return 21
        elif "aft" in x:
            return 14
        elif "eve" in x:
            return 21
        elif "mor" in x:
            return 9
        elif "mid" in x:
            return 12
        elif "dus" in x:
            return 22
        elif "dar" in x:
            return 22
        elif "noon" in x:
            return 12
        elif "sun" in x:
            return 12
        elif "dawn" in x:
            return 20
        elif "day" in x:
            return 8
        elif "a.m." in x:
            return 10
        elif "p.m." in x:
            return 18
        else:
            return x
    except:
        return x

def pares(x):
    try:
        matches = re.findall(r'(\d\d)', x)

        if matches and len(matches) >= 2:
            return matches[0] + "h" + matches[1]
        else:
            return x
    except:
        return x

def hour (time):
    try:
        if len(time) == 5:
            time = time.lower()
            time = time.split("h")

            hour = int(time[0])
            minut = int(time[1])

            if minut > 30:
                return int(hour) + 1
            elif minut <= 30:
                return int(hour)
            else:
                return "".join(time)
        else:
            return time
    except:
        return time

def clean(x):
    if isinstance(x, int):
        return x
    else:
        return 13

df.Time = df.Time.apply(pares)
df.Time = df.Time.apply(hour_letter)
df.Time = df.Time.apply(hour)
df.Time = df.Time.apply(clean)
df.Time = df.Time.apply(lambda x: 0 if x > 24 else x)

# Creación columna "Daytime"

def daytime (hora):
    if 6 <= hora < 12:
        return "Morning"
    elif 12 <= hora < 14:
        return "Noon"
    elif 14 <= hora < 20:
        return "Afternoon"
    else:
        return "Night"

df.Daytime = df.Time.apply(daytime)

# Creación columna "Season"

def season(month):
    if 1 <= month < 3:
        return "Winter"
    elif 3 <= month < 6:
        return "Spring"
    elif 6 <= month < 9:
        return "Summer"
    elif 9 <= month <= 12:
        return "Autumn"
        

df.Season = df.Month.apply(season)

df.Season.replace(np.NaN,"Unknown", inplace = True)

df.to_csv("../data/7_shark_time.csv",index=False)
