==> src/9_final.ipynb <==
import pandas as pd
import numpy as np
import geonamescache
import pycountry
import difflib
import warnings
import statistics as stat
import re

from countryinfo import CountryInfo
warnings.filterwarnings('ignore')
pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows',None)

df = pd.read_csv("../data/8_shark_species.csv")

# Limpieza final y comprobaciones

En este apartado vamos a realizar las últimas comprobaciones (valores nulos y casos atípicos). Además reiniciaremos el índice personal creado en la columna "Case Number" y prepararemos el data frame para una facil lectura.

Además sacaremos algunas estadísticas para tratar de obtener ciertas conclusiones.

df.head()

print(df.columns[df.isnull().any()].tolist())

Vemos que solo quedan valores nulos en la columna "Size". Recordamos que esto es una decisión tomada intencionalmente.

display(df.describe())

En la descripción general de los se detecta que el año máximo es 2176. A continuación corregimos el df para evitar la presencia de años por delante del actual.

display(df.loc[df['Year'] == 2176.00000])

df.loc[df['Year'] == 2176.00000, 'Year'] = 1959

Mientras se solucionaba el problema se detecto que únicamente había un año por encima de los valores posibles. Se decide cambiarlo a mano, encontrando el año del ataque en la columna "pdf".

display(df.describe())

Finalmente, y tras las últimas comprobaciones, declaramos la base de datos como limpia. Se ha eliminado la presencia de valores nulos, excepto los intencionadamente colocados en la columna Size y no se detectan más valores atípicos en la visualización.

# Estadísticas y conclusiones finales
